requirements.txt
fastapi[all]>=0.100.0<0.101.0
python-dotenv>=1.0.0,<1.1.0